compilation error fixed

+++ src/Socket/SSocket.cpp
@@ -6,7 +6,7 @@
 /*   By: agimi <agimi@student.1337.ma>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/03 10:02:05 by agimi             #+#    #+#             */
-/*   Updated: 2024/02/14 15:05:01 by agimi            ###   ########.fr       */
+/*   Updated: 2024/02/16 11:25:26 by agimi            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,7 +43,7 @@ sockaddr_in wbs::SSocket::get_add()
 	return add;
 }
 
-int wbs::SSocket::get_sfd()
+long wbs::SSocket::get_sfd()
 {
 	return sfd;
 }
